Fragment src/search.cpp
Moved all non-search critical functions outside of src/search.cpp
and into separate files. This reduces clutter, and makes this (most
important) file shorter and easier to comprehend (IMO).
